Add the link to the VS code terminal extension
The workshop participants shall install the VS code terminal extension.
Despite the fact that VS Code already has built-in support for the
terminal, this extension adds a terminal icon in the status bar at the
bottom. This is very handy as it makes it possible to reopen an already
existing closed terminal.